pyproject.toml
[tool.poetry]
name = "stock-news"
version = "0.1.0"
description = "Initially sends scheduled messages about news when triggered"
authors = [
    "Edson Pimenta <edson.tibo@gmail.com>",
    "Bruno Souza <bruno1912200@gmail.com>"
]
license = "MIT"
readme = "README.md"
packages = [{include = "stock_news"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^1.5.3"
requests = "^2.28.2"
types-requests = "^2.28.11.17"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"
pre-commit = "^3.2.1"
mypy = "^1.1.1"

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
